Add CSP to response headers
https://eaflood.atlassian.net/browse/PM-286

Following a PEN test it was flagged that the service has not specified a [Content Security Policy](https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP) in the response headers sent back from the app to a clients browser.

For reference CSP

> [..] is an HTTP response header that restricts the browser to loading external assets such as scripts, styles or media from a wide variety of sources — as well as inline scripts. It is intended to prevent wide categories of attacks, such as cross-site scripting (XSS), click-jacking and other forms of code injection.
>
> *https://blog.sqreen.io/integrating-content-security-policy-into-your-rails-applications-4f883eed8f45/*

Prior to this change we had logic that set custom headers via a `before_filter` in the `ApplicationController`. The issue found in testing was that only pages that originated in **pafs_core** had these headers applied. To ensure all pages from either the front or back office apps get the new CSP we first needed to refactor the setting of the headers to a module, and then update `ApplicationController` in **pafs_core** to set the response headers using the new module. The same change will be made to the [pafs-user](https://github.com/DEFRA/pafs-user) and [pafs-admin](https://github.com/DEFRA/pafs-admin) `ApplicationController`'s.

N.B. Ideally we would have liked to not set the policy for `script-src` to allow `unsafe-inline`, but currently the GOV.UK template makes use of inline JavaScript. CSP is fine with you allowing inline scripts, but ideally wants you to uniquely identify them using a nonce or sha. The team behind the template are aware of this though (See [Inline scripts need hashes provided for CSP directives](https://github.com/alphagov/govuk_template/issues/258) so hopefully once a solution is in place we can update our policy.