Add ability 2 generate file with multiple projects
https://eaflood.atlassian.net/browse/PM-297

This is the first step in offering users the ability to export all projects with a given status into a FCERM1 csv or spreadsheet. This change focuses on the back end, and expanding the exist functionality to

- search for all projects matching a given status for a user
- depending on user type this could encompass multiple areas
- export those projects into a xlsx file using the FCERM1 format

Users of the service have expressed that they still wish to be able to work with project data in spreadsheets as they currently do, but currently the functionality in the system only allows for the export of one project per file.

The functionality this change will add will be linked to changes in the UI in the future, that will allow users to access the new feature.