Add google analytics content 2 Content Sec. Policy
https://eaflood.atlassian.net/browse/PM-286

PR #128 was part of adding a Content-Security-Policy to the response header for all responses. However it was spotted that the layout in the **pafs-user** app has a section related to Google Analytics which is only added if the env var GOOGLE_ANALYTICS_ID is present.

We therefore needed to update the policy to include these sources from Google, as GOOGLE_ANALYTICS_ID will be present in the production environment.

The key changes are

- switch from us setting the Content-Security-Policy header to use [Secure Headers](ttps://github.com/twitter/secureheaders) gem
- update policy to cover content from Google to support Google analytics

The following covers why we moved away from setting the CSP header ourselves and instead brought in a gem.

------------------------------------------------------------------------

So we had updated the original implementation of adding a Content-Security-Policy header into a PAFS response, to handle Google Analytics. We then put our forward thinking hat on and considered what would happen when the [GOV.UK template is updated to work with CSP](https://github.com/alphagov/govuk_template/issues/258).

Currently when GOOGLE_ANALYTICS_ID is set we are adding a GA block into every page, part of which is an inline script. GA provides the script so this is as per their instructions. Once GOV.UK resolve the issue with their inline scripts and CSP, we'd be looking to remove 'unsafe-inline' from our policy, but doing so would break the GA script.

So we began by moving the code to a separate JavaScript file and pulling it through via the asset pipeline. The first issue was that we currently inject the GOOGLE_ANALYTICS_ID into the script using ERB. So we resolved this by creating a `google_analytics.js.erb` file so we could do the same in the new external file.

However this then threw up issues

- if the env var is not set we would get a JavaScript error reported on every page
- the asset fingerprinting feature does not recognise a change in the env var, so having pre-compiled the asset with the env var set one way we're stuck with it unless we manually delete pre-compiled version (there is a feature for this coming soon <https://github.com/rails/sprockets/blob/master/lib/sprockets/context.rb#L152> but its not available in the latest release)

So next step was to add a hidden field to the document which contained the ID, and amending the script to both grab the ID from the hidden field, and to abort if its not set. But that didn't work either!

Now the script is in a separate file and therefore being served by Rails asset pipeline, it was being added in the head. And as an immediately invoked function it was failing to find the hidden field because only the HEAD had been returned!

So this then led to a solution where we added a new meta tag to the head, defined before we add the scripts, and the script was amended to pull the ID from this. We now had a working solution however

- We had to add data into a meta tag, when meta tags are supposed to be data about data
- We'd moved the snippet substantially away from that supplied by Google, which means should it get updated in the future, we'd first have to review and rewrite whatever Google provided
- If the GOV.UK CSP solution still left the scripts inline, we still wouldn't have a solution for specifying agreed nonce values
- It felt like we'd gone very deep to solve what is a simple issue

So going back to the drawing board, and ideally looking for a solution that would work with Nonce values in expectation of a GOV.UK fix we hit upon [Secure Headers](https://github.com/twitter/secureheaders) from Twitter. This is a gem that you add to your Rails app, provide some configuration in an initializer, and it then automatically adds secure headers to all your responses.

Implementing this led to far less changes on the front end apps (simply need to add a line into `config.ru` as the 'automatic' bit didn't quit work - possibly because we actually set things up in an engine). Plus it would give us the ability to use nonce values in the views that would match what had been generated in the header (see https://github.com/twitter/secureheaders/blob/master/docs/per_action_configuration.md#nonce).

Obviously this meant completely refactoring how we had previously implemented the Content-Security-Header, so the changes required to switch to **Secure headers** were

- implementing Secure Headers including the config
- remove setting the Content-Security-Policy in `custom_headers.rb`
- refactoring `custom_headers.rb` to be more about non-security related custom headers
- refactoring `ApplicationController.rb` and associated tests to account for the changes