Update RFCC codes to include test areas
https://eaflood.atlassian.net/browse/PM-151

As part of finalising the areas.csv file which will be used to seed the database (see https://github.com/DEFRA/pafs-user/pull/116) we added in some test areas. This will allow the team to investigate issues, or demonstrate functionality without adding dummy data to real areas.

However we had not considered that we have RFCC codes which are used to determine the project reference number, that did not recognise these new test areas.

This change fixes that by adding a new RFC code (TS) which projects linked to test areas will use, and ensuring an validation will work with the new RFCC code.