Align areas in RFCC codes with latest areas
https://eaflood.atlassian.net/browse/PM-151

As part of updating the list of areas to be imported into the service for go live we also sanity checked the list of PSO areas used in RFCC codes against it.

We found there were some discrepencies in the names and even a missing record! This would have led to errors in the live system hence this change to align the PSO areas used in the code with those in the final version of the [areas.csv](https://github.com/DEFRA/pafs-user/blob/develop/lib/fixtures/areas.csv).