Add function and test to export multiple projects
https://eaflood.atlassian.net/browse/PM-297

Have added a test and a function that essentially copis the existing functionality for exporting a project to a spreadsheet, but builds on it by allowing for multiple projects to be exported.

The intention is to link up this feature with the changes to the `search()` method in `PafsCore::ProjectService` to allow all projects belonging to a user with the status of 'submitted' to be exported.

However currently it does not work (the test fails) because the xlsx tenplate we use has only a fixed number of rows. It was built for just exporting 1 project so when `add_project_to_sheet()` is called with a `row_no` greater than 7 it immediately breaks.

Either when adding a project we also need to automatically insert a new row to the spreadsheet. Or we need to amend the template to have enough rows to meet the largest possible export.

Once this was working, due to the duplication between `generate_xlsx()` and `generate_multi_xlsx()` the final step was going to be to refactor `generate_xlsx()` so it could handle either a single project, or a collection of them.