[PM-606] Use presigned URLs for file downloads (#304)
Currently, the app streams downloads from S3 and passes them onto the
client, which ties up app workers and the web server. We're also seeing
timeouts from the web server with large files.

This switches to using short-lived, presigned S3 URLs that hand the
download off so that it can be performed by the client directly to S3
without tying up the web or app server.